Allow admins to delete impersonation logs
Without this setting, admins are unable to delete Users
who have ever been impersonated.

This obviously opens up the possibility for a staff user
to impersonate a User, then "cover their tracks" by
deleting the log – but given this is a just an Alpha
right now, we’re not worried about that.