Teams can now have an optional "boundary"
Previously, teams only had a centroid point location (calculated from
the postcode of the team’s founder). Now teams can additionally have a
polygon boundary.

The boundary isn’t used for anything yet, and isn’t displayed anywhere
except in the Django Admin – but in the next commit, we’ll use Leaflet
to draw the team’s boundary on the Team page.

Team founders can’t yet define a boundary for their team via the
front-end, but site admins can add a boundary through the Django admin
(drawn by hand on the GeoDjango OSM map, or uploaded as a GeoJSON file).

I’ve included a GeoJSON file for a team in Marple in the fixtures
directory, and a new management command to create the team and import
the boundary from the file.