Display team boundary (and user postcode location) on Team page
The Leaflet code we previously used on the local authority "Area" page
has now been expanded to work on Team pages.

If the team has a boundary set, it is displayed on the map. Otherwise,
the map is just centred on the team’s centroid location.

If the user has arrived on the team page after searching for their
postcode on the Search page, then the location of that postcode is also
displayed on the Team map, to make it visually obvious whether you’re
within the team’s boundary or not (if it has one).

We use the OSM HOT map tiles because they’re less visually distracting
than the default "road atlas"-style OSM tiles, and the colours are
coincidentally a better match with the Neighbourhood Warmth branding.

The Leaflet.snogylop plugin dumped at the top of map.js allows us to
shade *outside* the boundary area (rather than the Leaflet default of
only being able to shade inside polygons) making it stand out better.

The user’s last searched-for location is stored in their session.
There is currently no way to clear it (other than deleting the django
session cookie).