Make team progress appear on public team page
An unnecessary `else` in `views.py` was causing custom challenge
progress not to be shown on the team page if a user was logged in.
And the challenge names/short_descriptions weren’t being displayed
properly on the progress bar either.

I‘ve also renamed the `challenges` environment variable to
`public_challenges` to make it clear that it doesn’t include the
full list of the team’s current challenges.